Github.readme
I am an eccentric guy with a passion for game programming and development. I have a small group of friends/acquaintances helping me already via a server I created and manage on discord in the area of concept art. With that said, this idea is inspired by games like Skyrim, Diablo III, and DnD; with a ton of its own personal tweaks and changes.